refactor: move functions used only in orb_id_linux.rs there.
