+++ orb-info/src/lib.rs
@@ -13,7 +13,6 @@ pub mod orb_token;
 
 use std::io;
 use std::path::Path;
-use std::process::Output;
 
 #[cfg(feature = "orb-id")]
 pub use orb_id::OrbId;
@@ -41,27 +40,4 @@ async fn from_file(path: impl AsRef<Path>) -> io::Result<String> {
 )]
 fn from_file_blocking(path: impl AsRef<Path>) -> io::Result<String> {
     std::fs::read_to_string(path).map(|s| s.trim().to_owned())
-}
-
-#[cfg(feature = "async")]
-async fn from_binary(path: impl AsRef<Path>) -> io::Result<String> {
-    let output = tokio::process::Command::new(path.as_ref()).output().await?;
-    from_binary_output(output, path)
-}
-
-#[cfg_attr(not(feature = "orb-id"), expect(dead_code))]
-fn from_binary_blocking(path: impl AsRef<Path>) -> io::Result<String> {
-    let output = std::process::Command::new(path.as_ref()).output()?;
-    from_binary_output(output, path)
-}
-
-fn from_binary_output(output: Output, path: impl AsRef<Path>) -> io::Result<String> {
-    if output.status.success() {
-        Ok(String::from_utf8_lossy(&output.stdout).to_string())
-    } else {
-        Err(std::io::Error::other(format!(
-            "{} binary failed",
-            path.as_ref().display()
-        )))
-    }
 }
